rename files to fix tests
