+++ deeplabcut/pose_estimation_pytorch/apis/__init__.py
@@ -15,12 +15,12 @@ from deeplabcut.pose_estimation_pytorch.apis.analyze_images import (
     analyze_image_folder,
     superanimal_analyze_images,
 )
-from deeplabcut.pose_estimation_pytorch.apis.analyze_videos import (
+from deeplabcut.pose_estimation_pytorch.apis.videos import (
     analyze_videos,
     video_inference,
     VideoIterator,
 )
-from deeplabcut.pose_estimation_pytorch.apis.convert_detections_to_tracklets import (
+from deeplabcut.pose_estimation_pytorch.apis.tracklets import (
     convert_detections2tracklets,
 )
 from deeplabcut.pose_estimation_pytorch.apis.evaluation import (
